Fixing the formatting of files under the branch 091222
