+++ 091222/code/transaction_log.txt
@@ -1,0 +1,38 @@
+2025-03-08 14:00:20 12345679 560000.00 DEPOSIT 560000.00
+2025-03-08 14:00:53 12345679 12433.00 WITHDRAW 547567.00
+2025-03-08 14:04:19 12345680 82383920.00 DEPOSIT 82383920.00
+2025-03-08 14:04:23 12345680 12324.00 WITHDRAW 82371600.00
+2025-03-09 10:02:19 12345679 121221.00 DEPOSIT 121221.00
+2025-03-09 10:03:08 12345679 121221.00 WITHDRAW 0.00
+2025-03-09 10:03:15 12345679 1200.00 DEPOSIT 1200.00
+2025-03-09 10:03:21 12345679 600.00 WITHDRAW 600.00
+2025-03-09 10:49:52 12345679 2323.00 DEPOSIT 2323.00
+2025-03-09 11:16:29 12345679 2121.00 DEPOSIT 2121.00
+2025-03-09 11:16:38 12345679 21.00 WITHDRAW 2100.00
+2025-03-09 11:18:18 12345679 12121.00 DEPOSIT 14221.00
+2025-03-09 11:20:57 12345679 120000.00 DEPOSIT 120000.00
+2025-03-09 11:21:03 12345679 121.00 WITHDRAW 119879.00
+2025-03-09 11:50:39 12345682 100000002537764290000000000.00 DEPOSIT 100000002537764290000000000.00
+2025-03-09 11:50:48 12345682 99999.00 DEPOSIT 100000002537764290000000000.00
+2025-03-09 11:50:58 12345682 9999999827968.00 WITHDRAW 100000002537764290000000000.00
+2025-03-09 11:51:12 12345682 10000000272564224.00 WITHDRAW 100000002537764290000000000.00
+2025-03-09 22:41:18 12345683 30000000.00 DEPOSIT 30000000.00
+2025-03-09 22:41:26 12345683 234234.00 WITHDRAW 29765766.00
+2025-03-09 22:41:36 12345683 356536.00 WITHDRAW 29409230.00
+2025-03-09 23:23:42 12345679 122.00 WITHDRAW 119757.00
+2025-03-10 15:17:29 12345679 332.00 DEPOSIT 120089.00
+2025-03-10 15:19:45 12345679 232.00 WITHDRAW 119857.00
+2025-03-10 16:35:29 12345680 12121.00 DEPOSIT 12121.00
+2025-03-10 16:35:36 12345680 12.00 WITHDRAW 12109.00
+2025-03-10 17:07:47 12345679 1200.00 TRANSFER 118657.00
+2025-03-10 17:07:47 12345680 1200.00 RECEIVE 1200.00
+2025-03-10 17:08:14 12345679 1200.00 TRANSFER 117457.00
+2025-03-10 17:08:14 12345681 1200.00 RECEIVE 1200.00
+2025-03-10 17:08:55 12345679 999.00 TRANSFER 116458.00
+2025-03-10 17:08:55 12345680 999.00 RECEIVE 2199.00
+2025-03-10 17:09:29 12345679 1200.00 TRANSFER 115258.00
+2025-03-10 17:09:29 12345683 1200.00 RECEIVE 29410430.00
+2025-03-15 17:59:21 12345679 12.00 DEPOSIT 115270.00
+2025-03-15 17:59:28 12345679 12.00 WITHDRAW 115258.00
+2025-03-15 18:01:07 12345679 20000.00 TRANSFER 95258.00
+2025-03-15 18:01:07 12345680 20000.00 RECEIVE 22199.00
