--- code/userdetail.txt
@@ -1,6 +1,0 @@
-12345679 nirajan 9812606901 Chaudhary@1221 115258.00 2020-12-12 kathmandu nirajanchaudhary@gmail.com
-12345680 prajwal 9834343434 Fack@1221 2199.00 2063-12-21 bardiya thru@gmail.com
-12345681 bhumika 9812606902 Bhumi@1221 1200.00 2056-12-21 kathmandu Bhumi@gmail.com
-12345682 prajwal 9828230477 Kalebhale@79 100000002537764290000000000.00 2063-06-31 bardiya prajwaltharu79@gmail.com
-12345683 bidhan 9841000000 Bidhan@123 29410430.00 2082-03-10 mahalaxmi sthabdn@hotmail.com
-12345684 bidhan 9864156204 Bidhan@2023 0.00 2025-01-10 imadol sthabdn@gmail.com
